refactor: Added abstraction for PreferencesDataSource
