+++ personalization-sdk/src/main/kotlin/com/personalization/di/SdkComponent.kt
@@ -2,6 +2,7 @@ package com.personalization.di
 
 import com.personalization.SDK
 import com.personalization.features.notification.service.NotificationService
+import com.personalization.sdk.data.di.AbstractDataSourcesModule
 import com.personalization.sdk.data.di.DataSourcesModule
 import com.personalization.sdk.data.di.ModelsModule
 import com.personalization.sdk.data.di.RepositoriesModule
@@ -14,6 +15,7 @@ import javax.inject.Singleton
         DataSourcesModule::class,
         RepositoriesModule::class,
         ModelsModule::class,
+        AbstractDataSourcesModule::class,
         SdkModule::class,
         AppModule::class
     ]
@@ -30,5 +32,4 @@ interface SdkComponent {
     fun inject(sdk: SDK)
 
     fun inject(service: NotificationService)
-
 }
